feat: move scraping logic into class and add tkinter gui
The output of the gui has an issue where it will only display messages
sent to stdout *after* the scraping is finished. This is caused by the
scraping function blocking the gui's main loop from updating. We can't
run the scraping function in a separate thread because pyppeteer uses
python's signal module, which throws an error if it is used outside the
main thread. Pyppeteer is used by requests-html under the hood, so we
should switch from requests-html to playwright.